requirements.txt
aldryn-apphooks-config==0.3.3
aldryn-boilerplates==0.7.5
aldryn-common==1.0.4
aldryn-search==0.3.0
Babel==2.3.4
beautifulsoup4==4.6.0
boto==2.46.1
chardet==3.0.4
cmsplugin-filer==1.1.3
codegen==1.0
coverage==4.2
defusedxml==0.4.1
Django==1.9.11
django-allauth==0.28.0
django-appconf==1.0.2
django-appdata==0.1.6
django-axes==2.3.2
django-blog-it==0.2.3
-e git://github.com/Fantomas42/django-blog-zinnia.git@bdeb44977c6e23d89d77ef47f505ced5f133ff68#egg=django_blog_zinnia
django-bower==5.2.0
django-braces==1.11.0
django-categories-i18n==1.0
django-ckeditor==5.3.0
django-classy-tags==0.8.0
django-cms==3.4.4
django-compressor==2.1
django-contrib-comments==1.8.0
django-easy-timezones==0.8.0
django-extra-views==0.6.4
django-facebook==6.0.3
django-filer==1.2.7
django-fluent-blogs==1.2.4
django-fluent-contents==1.2.1
django-fluent-utils==1.3.2
django-formtools==2.0
django-haystack==2.5.1
django-jenkins==0.110.0
django-js-asset==0.1.1
django-localflavor==1.3
django-markdown==0.8.4
django-meta==1.3.2
django-meta-mixin==0.3.0
django-modelcluster==3.1
django-mptt==0.8.7
django-nose==1.4.4
django-oscar==1.3
django-oscar-datacash==0.8.3
django-oscar-paypal==0.9.7
django-parler==1.8
django-polymorphic==1.0.2
django-sekizai==0.10.0
django-ses==0.8.2
django-simple-captcha==0.5.5
django-simple-pagination==1.1.8
django-slug-preview==1.0.2
django-smtp-ssl==1.0
django-sortedm2m==1.4.0
django-spurl==0.6.4
django-standard-form==1.1.1
django-storages==1.6.5
django-tables2==1.0.7
django-tag-parser==3.1
django-tagging==0.4.5
django-taggit==0.22.1
django-template-analyzer==1.6.1
django-templatetag-sugar==1.0
django-treebeard==4.0.1
django-webpack-loader==0.3.3
django-widget-tweaks==1.4.1
django-wysiwyg==0.8.0
django-xmlrpc==0.1.7
djangocms-admin-style==1.2.7
djangocms-apphook-setup==0.1.2
djangocms-attributes-field==0.3.0
djangocms-blog==0.8.12
djangocms-text-ckeditor==3.5.0
djangorestframework==3.6.3
dsdev-utils==0.9.6
easy-thumbnails==2.4.1
factory-boy==2.7.0
fake-factory==0.7.2
future==0.16.0
gunicorn==19.6.0
html5lib==0.9999999
httplib2==0.9.2
ipaddress==1.0.16
lxml==3.8.0
Markdown==2.6.8
Menus==0.2.0
microurl==0.1.1
mock==2.0.0
mots-vides==2015.5.11
mysqlclient==1.3.9
nose==1.3.7
oauth2==1.9.0.post1
oauthlib==2.0.0
olefile==0.44
patterns==0.3
pbr==1.10.0
phonenumbers==7.7.3
Pillow==4.2.1
pluggy==0.4.0
purl==1.3
py==1.4.33
pycountry==1.20
pygeoip==0.3.2
PyJWT==1.4.2
pyparsing==2.2.0
pysolr==3.6.0
pytest==3.0.7
pytest-django==3.1.2
python-dateutil==2.5.3
python-social-auth==0.2.21
python3-openid==3.0.10
pytz==2017.2
rcssmin==1.0.6
regex==2017.7.11
requests==2.11.1
requests-oauthlib==0.6.1
requirements==0.1
rjsmin==1.0.12
selenium==3.0.2
six==1.10.0
sorl-thumbnail==12.4a1
South==1.0.2
stripe==1.48.1
tox==2.6.0
Unidecode==0.4.19
URLObject==2.4.3
virtualenv==15.1.0
Whoosh==2.7.4
Willow==0.4
YURL==0.13
